Move index iterator into own module
