+++ non_contiguously_indexed_array/src/lib.rs
@@ -3,3 +3,6 @@ pub use array::*;
 
 mod index;
 pub use index::*;
+
+mod iter;
+use iter::*;
